+++ ocdskit/cli/commands/compile.py
@@ -4,7 +4,7 @@ import sys
 import ocdskit.packager
 from ocdskit.cli.commands.base import OCDSCommand
 from ocdskit.combine import merge
-from ocdskit.exceptions import CommandError, InconsistentVersionError
+from ocdskit.exceptions import CommandError, InconsistentVersionError, MissingOcidKeyError
 
 logger = logging.getLogger('ocdskit')
 
@@ -39,6 +39,8 @@ class Command(OCDSCommand):
         try:
             for output in merge(self.items(), streaming=True, **kwargs):
                 self.print(output, streaming=self.args.package)
+        except MissingOcidKeyError:
+            raise CommandError('The `ocid` field of at least one release is missing.')
         except InconsistentVersionError as e:
             versions = [e.earlier_version, e.current_version]
             if versions[1] < versions[0]: